onboarding_guide_agent/requirements.txt
openai
python-dotenv
chromadb
fastapi
uvicorn
pydantic
tenacity
sqlalchemy
aiosqlite
asyncpg
aiomysql
sqlalchemy